[dev.ssa] cmd/compile: teach register allocator about in-place ops
DO NOT SUBMIT

Help me, khr-wan-kenobi.

This does not work yet. I think I must have
the regalloc change wrong, but I don't see why.

The error I get is:

localOffset of non-LocalSlot value: v16 = XORQconst <bool> [1] v6 : CX

which doesn't make any sense?!

Also, if we do this, we should go ahead and mark
the other in-place ops as such.

CONTEXT:

This manifested itself as a bug in the mime package
generated code.

The problem is that the register allocator doesn't
know that XORQconst is an in-place operation.

Here's the function:

func isNotTokenChar(r rune) bool {
  return !isTokenChar(r)
}

At the end of stackalloc, here's the SSA (explanation follows):

isNotTokenChar <nil>
  b1:
    v1 = Arg <mem>
    v2 = SP <uint64> : SP
    v11 = LEAQ <*rune> {r} v2 : AX
    v17 = LEAQ <*bool> {~r1} v2 : CX
    v19 = MOVLload <rune> v11 v1 : DX
    v10 = MOVLconst <int8> : BX
    v9 = MOVBstore <mem> v17 v10 v1
    v12 = MOVLstore <mem> v2 v19 v9
    v8 = ADDQconst <*bool> [8] v2 : BP
    v5 = StoreReg <*bool> v17 : 16(SP)
    v14 = StoreReg <*bool> [8] v8 : 24(SP)
    v13 = CALLstatic <mem> [16] {isTokenChar} v12
    Call v13 -> b3 b2
  b3: <- b1
    v7 = LoadReg <*bool> v14 : AX
    v6 = MOVBload <bool> v7 v13 : CX
    v3 = XORQconst <bool> [1] v6 : DX
    v23 = LoadReg <*bool> v5 : BX
    v18 = MOVBstore <mem> v23 v3 v12
    Plain -> b2
  b2: <- b1 b3 b4
    v22 = Phi <mem> v12 v18 v1
    Exit v22

The interesting bit is b3.
We load the result of the function call into CX (v6).
We xor it with 1 to negate it and put the result in DX (v3)
(except the result is actually in CX still!)
We store that back to our stack slot (v18).

It generates this assembly snippet:

MOVQ 24(SP), AX  // get address of result
MOVB (AX), CL    // load result into CX
XORQ $1, CX      // not result, in-place
MOVQ 16(SP), BX  // get addr of ret val
MOVB DL, (BX)    // write ret val, *wrong*
                 // should be CL instead of DL, DX contains junk

Change-Id: I759422e1e80ff16c98a8488283b7ffc20af76570