[dev.ssa] cmd/compile: catch known genssa errors earlier
DO NOT SUBMIT

This is a hack to enable testing more generated code.
I mailed it so that others can easily use it.

With this hack in place, run:

for pkg in `go list std`; do GOGC=off GOSSAPKG=`basename $pkg` go test $pkg; done

As of commit ca088cf4, this results in:

* 130 passing tests
* some tests that fail due to GOGC=off:
  runtime, runtime/debug, runtime/pprof, runtime/trace, sync
  (perhaps we should detect GOGC=off in the relevant tests and t.Skip?)
* fmt tests fail due to issue #12054
* math/big tests fail, undiagnosed (but I'm about to work on it)

Change-Id: Ia770d5c1a37d89f5c57ca70a93fd5285a0fc66bb